# Refactor favicon middleware, avoid magic numbers.