Increase readability / follow best practices.