@vikpe 2022-05-28
Update: I introduced a constant for the standard path and ran the benchmark.

**master**
```
cpu: Intel(R) Core(TM) i5-4670K CPU @ 3.40GHz
Benchmark_Middleware_Favicon
go test -v -bench=Benchmark_Middleware_Favicon -benchmem -count=4
Benchmark_Middleware_Favicon-4          13465311                86.32 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          13408497                87.49 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          13480440                87.96 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          12733093                87.11 ns/op            3 B/op          1 allocs/op
```

**pr**
```
cpu: Intel(R) Core(TM) i5-4670K CPU @ 3.40GHz
Benchmark_Middleware_Favicon
Benchmark_Middleware_Favicon-4          13345022                86.64 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          13459353                86.65 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          13541000                86.20 ns/op            3 B/op          1 allocs/op
Benchmark_Middleware_Favicon-4          13520474                86.09 ns/op            3 B/op          1 allocs/op
```

However I guess one need to run benchmarks on a dedicated setup (not my dev computer) for persistent results.